reject incremental cursors and join projections on incomplete columns
