+++ tests/dataset/conftest.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from typing import Any, Iterator, TYPE_CHECKING
 
 import dlt
 import pytest
@@ -97,4 +97,51 @@ def dataset_with_annotated_references(module_tmp_path: pathlib.Path) -> dlt.Data
 
     pipeline.run(annotated_references())
 
+    return pipeline.dataset()
+
+
+@pytest.fixture(scope="module")
+def dataset_with_incomplete_join_target(module_tmp_path: pathlib.Path) -> dlt.Dataset:
+    """Two sibling tables joined by an explicit reference, where the join target
+    declares an incomplete column hint via `columns=`.
+
+    `phantom_field` is declared on `categories` with no `data_type`, so it never
+    materializes at the destination. `Schema.get_table_columns()` filters it out
+    via `is_complete_column`; raw `schema.tables[...]["columns"]` does not.
+    """
+    pipeline = dlt.pipeline(
+        pipeline_name="relation_incomplete_join_target",
+        pipelines_dir=str(module_tmp_path / "pipelines_dir_incomplete"),
+        destination=dlt.destinations.duckdb(str(module_tmp_path / "incomplete.db")),
+        dev_mode=True,
+    )
+
+    @dlt.resource(
+        name="categories",
+        primary_key="id",
+        columns=[{"name": "phantom_field", "nullable": True}],
+    )
+    def categories() -> Iterator[Any]:
+        yield [{"id": 1, "name": "alpha"}, {"id": 2, "name": "beta"}]
+
+    @dlt.resource(
+        name="products",
+        primary_key="id",
+        columns=[{"name": "category_id", "data_type": "bigint"}],
+        references=[
+            {
+                "referenced_table": "categories",
+                "columns": ["category_id"],
+                "referenced_columns": ["id"],
+            }
+        ],
+    )
+    def products() -> Iterator[Any]:
+        yield [
+            {"id": 10, "category_id": 1},
+            {"id": 11, "category_id": 2},
+            {"id": 12, "category_id": 1},
+        ]
+
+    pipeline.run([categories(), products()])
     return pipeline.dataset()
